examples/basic-nextjs-fastapi/backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0
gittielabs-nextjs-fastapi-auth>=0.1.0
